Move the map out of App.js and in to MapScren.js. Install react-navigation npm module and add stacknavigator to App.js
