+++ package.json
@@ -9,7 +9,8 @@
 	"dependencies": {
 		"react": "16.0.0-alpha.12",
 		"react-native": "^0.48.3",
-		"react-native-maps": "^0.16.3"
+		"react-native-maps": "^0.16.3",
+		"react-navigation": "^1.0.0-beta.11"
 	},
 	"devDependencies": {
 		"babel-jest": "21.0.2",
