+++ src/components/MapScreen.js
@@ -1,0 +1,91 @@
+import React, { Component } from 'react';
+import {
+  StyleSheet,
+  View
+} from 'react-native';
+
+import MapView, { PROVIDER_GOOGLE } from 'react-native-maps';
+
+class ReactMaps extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      coords: []
+    };
+  }
+
+  componentDidMount() {
+    // find your origin and destination point coordinates and pass it to our method.
+    // I am using Bursa,TR -> Istanbul,TR for this example
+    this.getDirections("40.1884979, 29.061018", "41.0082,28.9784")
+  }
+   async getDirections(startLoc, destinationLoc) {
+    try {
+        let resp = await fetch(`https://maps.googleapis.com/maps/api/directions/json?origin=${ startLoc }&destination=${ destinationLoc }`)
+        let respJson = await resp.json();
+        let points = Polyline.decode(respJson.routes[0].overview_polyline.points);
+        let coords = points.map((point, index) => {
+            return  {
+                latitude: 37.78825,
+                longitude: -122.4324,
+                latitudeDelta: 0.015,
+                longitudeDelta: 0.0121,
+            }
+        })
+        this.setState({coords: coords})
+        return coords
+    } catch(error) {
+        return error
+    }
+}
+
+  render() {
+    return (
+      <View style={styles.container}>
+
+        <MapView
+          provider={PROVIDER_GOOGLE}
+          style={styles.map}
+          zoomEnabled={true}
+
+          initialRegion={{
+            latitude: 137.78825,
+            longitude: -122.4324,
+            latitudeDelta: 0.015,
+            longitudeDelta: 0.0121,
+          }}
+          />
+
+          <MapView.Polyline
+            coordinates={this.state.coords}
+            strokeWidth={2}
+            strokeColor="red"/>
+
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#F5FCFF',
+  },
+  welcome: {
+    fontSize: 20,
+    textAlign: 'center',
+    margin: 10,
+  },
+  instructions: {
+    textAlign: 'center',
+    color: '#333333',
+    marginBottom: 5,
+  },
+  map: {
+    ...StyleSheet.absoluteFillObject,
+  },
+});
+
+export default ReactMaps;
